# Cut dependencies for building ejabberd