--- a/Dockerfile
+++ b/Dockerfile
@@ -1,27 +1,26 @@
 FROM centos:centos6
 
 RUN yum -y update --exclude=filesystem && \
-    yum -y install epel-release && \
     yum -y install \
         gcc \
         gcc-c++ \
         rpmdevtools \
         tar \
-        yum-utils \
-    && \
-    yum clean all
+        yum-utils
 
 COPY erlang_solutions.asc /etc/pki/rpm-gpg/
 COPY erlang_solutions.repo /etc/yum.repos.d/
 RUN rpm --import /etc/pki/rpm-gpg/erlang_solutions.asc
-
-RUN useradd -m -s /bin/bash build
 
 WORKDIR /home/build
 RUN mkdir -p rpmbuild/{BUILD,BUILDROOT,RPMS,SOURCES,SPECS,SRPMS}
 COPY ejabberd.spec rpmbuild/SPECS/
 COPY sources rpmbuild/SOURCES/
 COPY build.sh ./
-RUN chown -R build:build .
+RUN useradd -m -s /bin/bash build && \
+    chown -R build:build .
+
+RUN spectool -C rpmbuild/SOURCES/ -g rpmbuild/SPECS/ejabberd.spec
+RUN sed -n 's:^BuildRequires\: *\([^ ]*\).*:\1:p' rpmbuild/SPECS/ejabberd.spec | xargs yum -y install
 
 CMD sh build.sh