--- a/Dockerfile
+++ b/Dockerfile
@@ -12,13 +12,13 @@ COPY erlang_solutions.asc /etc/pki/rpm-gpg/
 COPY erlang_solutions.repo /etc/yum.repos.d/
 RUN rpm --import /etc/pki/rpm-gpg/erlang_solutions.asc
 
+RUN useradd -m -s /bin/bash build
 WORKDIR /home/build
 RUN mkdir -p rpmbuild/{BUILD,BUILDROOT,RPMS,SOURCES,SPECS,SRPMS}
 COPY ejabberd.spec rpmbuild/SPECS/
 COPY sources rpmbuild/SOURCES/
 COPY build.sh ./
-RUN useradd -m -s /bin/bash build && \
-    chown -R build:build .
+RUN chown -R build:build .
 
 RUN spectool -C rpmbuild/SOURCES/ -g rpmbuild/SPECS/ejabberd.spec
 RUN sed -n 's:^BuildRequires\: *\([^ ]*\).*:\1:p' rpmbuild/SPECS/ejabberd.spec | xargs yum -y install