--- a/circle.yml
+++ b/circle.yml
@@ -8,7 +8,7 @@ dependencies:
   pre:
     - pip install https://github.com/abusesa/github-release/archive/master.tar.gz
   override:
-    - docker build --pull -t build-image .
+    - docker build --pull --rm -t build-image .
 
 test:
   override: