Items in this PR:

 * Do not require full Erlang for building. The erlang package seems to be some kind of meta package with big list of dependencies. Include only what is necessary for build. Speeds up build quite a bit.
 * Do more in docker container build. Try to get as much as possible of the CentOS packages installed. This mostly benefits debugging and devel builds.
 * build docker image with `--rm` to remove the intermediate containers.
 * Remove OpenSSL version dependency from ejabberd build.