@oherrala 2016-04-12
Small size and time comparison. Times are average of three runs. For build, the images (except centos6) were cleaned.

Before diet:

 * docker build: 397 MB, ~56 seconds
 * docker run: 407.4 MB (virtual 804.4 MB), ~218 seconds

After diet:

 * docker build:  547 MB, ~ 81 seconds
 * docker run: 16.39 MB (virtual 563.2 MB), ~108 seconds

So, before diet the build+run took on average 274 seconds and after diet 189 seconds. Speedup of 85 seconds. And disk required dropped about 241 MB.